Create Source Code
This is mostly in python. It still needs alot of work. I'd like to get the program to the point where it asks you for the information it needs to begin building out your budget and then returns the remaining balance to you once you've accounted for all factors. Will continue to work on it as I learn.